Define ImTextureID to VkDescriptorSet
ImTextureID looks like this in imgui.h:

// ImTexture: user data for renderer backend to identify a texture [Compile-time configurable type]
// - To use something else than an opaque void* pointer: override with e.g. '#define ImTextureID MyTextureType*' in your imconfig.h file.
// - This can be whatever to you want it to be! read the FAQ about ImTextureID for details.
#ifndef ImTextureID
typedef void* ImTextureID;          // Default: store a pointer or an integer fitting in a pointer (most renderer backends are ok with that)
#endif

Just define it to VkDescriptorSet in imconfig.h instead of casting to
void *.

Looked into this now to avoid having to include imgui.h down in Vulkan
backend header of renderer. Feels a bit hackish. But then again the whole
ImGui integration feels a bit hackish. This is nicer than using a raw
void *... probably... I think... maybe. Hack hack hack.