Define ImTextureID to VkDescriptorSet
ImTextureID looks like this in imgui.h:

// ImTexture: user data for renderer backend to identify a texture [Compile-time configurable type]
// - To use something else than an opaque void* pointer: override with e.g. '#define ImTextureID MyTextureType*' in your imconfig.h file.
// - This can be whatever to you want it to be! read the FAQ about ImTextureID for details.
typedef void* ImTextureID;          // Default: store a pointer or an integer fitting in a pointer (most renderer backends are ok with that)

Just define it to VkDescriptorSet in imconfig.h instead of casting to
void *.

Looked into this now to avoid having to include imgui.h down in Vulkan
backend header of renderer. Feels a bit hackish. But then again the whole
ImGui integration feels a bit hackish. This is nicer than using a raw
void *... probably... I think... maybe. Hack hack hack.

Could add IMGUI_USER_CONFIG define to project to point to user_imconfig.h
or something in directory above imgui directory. But want to avoid global
define in compiler command line. Should perhaps reconsider now that this
is only patch that is needed after Vulkan image support was merged. Also
nice to be able to use a completely clean upstream and not have a fork.