Do not set VkDescriptorSetLayoutBinding::pImmutableSamplers
From https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkDescriptorSetLayoutBinding.html :

"pImmutableSamplers affects initialization of samplers. If descriptorType
specifies a VK_DESCRIPTOR_TYPE_SAMPLER or
VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER type descriptor, then
pImmutableSamplers can be used to initialize a set of immutable samplers.
Immutable samplers are permanently bound into the set layout and must not
be changed; updating a VK_DESCRIPTOR_TYPE_SAMPLER descriptor with
immutable samplers is not allowed and updates to a
VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER descriptor with immutable
samplers does not modify the samplers (the image views are updated, but
the sampler updates are ignored)."

So setting it means the same sampler will be used for all textures.

One instance was removed in original pr (914) that added image support but
it got lost when it was merged. (Not sure why there are two
vkCreateDescriptorSetLayout() for bd->DescriptorSetLayout.)

Upstream issue # is 5502.